refactor test_training_integration to test_clis.py
This test uses different data from the other tests in this
module. The "indirect=" is a red flag. UserConfig is supposed to be a
user facing configuration, so it makes sense that a command line test
would specifiy this directly as an input.

Now the data for this test can freely be changed without breaking the
other tests.
